backend/app/requirements.txt
fastapi==0.78.0
uvicorn==0.12.2
python-dotenv==0.20.0
sqlalchemy==1.4.39
psycopg2==2.9.3
sqladmin==0.1.11
alembic==1.8.0
pytest==7.1.2
bcrypt==3.2.0
passlib==1.7.4
python-jose==3.3.0
python-frontmatter==1.0.0
Markdown==3.4.1
